refactor(web): extract registration form into RegistrationModal component
apps/web/app/profile/page.tsx was 529 lines mixing profile display,
verification state, and the entire registration dialog (form state,
validation, submit handler, transaction-pending flow) in one
component. Extract the dialog into components/profile/RegistrationModal.tsx
with its own props interface (isOpen, onClose, register, isRegistering,
registerError), taking the register mutation from useProfile as props
rather than calling the hook a second time.

Preserved behavior exactly, including a pre-existing bug: the
'Company name is required' check can never pass because there's no
Company Name input anywhere in the form (companyName is a useState
with no setter). Filed as a separate follow-up rather than fixed here
since adding the missing field is a product decision, not a mechanical
extraction — the new RegistrationModal.test.tsx documents today's
actual (broken) behavior so a future fix has a regression test to
update.

page.tsx: 529 -> 273 lines. New component: 285 lines, 285-line test
file with 7 cases (open/closed, close paths, validation, error
display, disabled state, role toggle). pnpm typecheck and pnpm
--filter web test both pass; web coverage unaffected (still >70%).
